Update copyright headers in files
Updated the copyright headers to use the SPDX ID for the license
to reduce the prefixing comment size.  Also cleaned up the rest
of the comment to be much more concise.

Finally, I noticed that the old LICENSE.txt file was still around.
Removed it, updated the LICENSE file with the comment that other
files in the repo contain non-Apache 2.0 licensed material and
removed the COPYRIGHT.txt file.